Modified for-loops that loop on stl containers to use size_t instead of int for the loop variable, in order to remove signed/unsigned integer comparison compiler warnings
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@4160 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/Exceptions/Database/InvalidFilesException.C
@@ -110,7 +110,7 @@ InvalidFilesException::InvalidFilesException(const char *filename,
     sprintf(str, "There was an error opening %s. It may be an "
             "invalid file.  VisIt tried using the following file format "
             "readers to open the file: ", filename);
-    for (int i = 0 ; i < plugins.size() ; i++)
+    for (size_t i = 0 ; i < plugins.size() ; i++)
     {
         if (i != plugins.size()-1)
             sprintf(strlen(str)+str, "%s, ", plugins[i].c_str());
@@ -147,7 +147,7 @@ InvalidFilesException::InvalidFilesException(const char *filename,
     sprintf(str, "There was an error opening %s. It may be an "
             "invalid file.  VisIt tried using the following file format "
             "readers to open the file: ", filename);
-    for (int i = 0 ; i < plugins.size() ; i++)
+    for (size_t i = 0 ; i < plugins.size() ; i++)
     {
         if (i != plugins.size()-1)
             sprintf(strlen(str)+str, "%s, ", plugins[i].c_str());
